Commit #refactor: refactor classes
