+++ src/com/company/Main.java
@@ -1,15 +1,10 @@
 package com.company;
-import com.company.model.Player;
 
-//import static com.company.controller.CharacterController.createCharacter;
-import static com.company.controller.CharacterController.createPlayer;
-//import static com.company.view.IOView.mainLoopView;
 import static com.company.view.IOView.mainLoopView;
 
 public class Main {
 
     public static void main(String[] args) {
-
         mainLoopView();
 
     }
