+++ src/com/company/service/FightService.java
@@ -73,13 +73,24 @@ public class FightService {
         }
     }
 
+//    public static boolean attackSuccess(Character attacker, Character defender) {
+//        Random r = new Random();
+//        int probability = r.nextInt(15 - 1) + 1;
+//        int result;
+//        if(attacker.getDexterity() == defender.getDexterity()){
+//            result = 5;
+//        }else result = attacker.getDexterity() - defender.getDexterity();
+//        return (result >= probability);
+//    }
+
     public static boolean attackSuccess(Character attacker, Character defender) {
         Random r = new Random();
-        int probability = r.nextInt(15 - 1) + 1;
-        int result;
-        if(attacker.getDexterity() == defender.getDexterity()){
-            result = 5;
-        }else result = attacker.getDexterity() - defender.getDexterity();
+        int dexterityDif = attacker.getDexterity() - defender.getDexterity();
+        double maxValue = 15 - dexterityDif / 2;
+        double probability;
+        if (maxValue <= 0) probability = r.nextDouble(1 - maxValue) + maxValue;
+        else probability = r.nextDouble(maxValue - 1) + 1;
+        double result = 7.5 + dexterityDif / 2;
         return (result >= probability);
     }
 
@@ -94,6 +105,11 @@ public class FightService {
         player.setDefense(player.getDefense() + 5);
         player.setSpeed(player.getSpeed() + 5);
         player.setSpeed(player.getDexterity() + 5);
+    }
+
+    public static int valueGained(){
+        Random r = new Random();
+        return r.nextInt(5-1)+1;
     }
 
 }
